requirements.txt
streamlit
pandas
opencv-python
openai
llama-index
tiktoken 
unstructured
chromadb
langchain 
google-search-results
wikipedia
pydantic 
sqlalchemy
python-dotenv